+++ tests/mirror/jupyter_again.py
@@ -26,9 +26,7 @@ editor_options:
      chunk_output_type console
 '''
 
-
 import yaml
 print(yaml.dump(yaml.load(c)))
-
 
 ?next